refactor(@types): centralize interfaces and types
- delete @types/statusCodes and @types/validation
- correct affected files
