--- src/@types/statusCodes/ISuccess.ts
@@ -1,13 +1,0 @@
-type SuccessStatusCode =
-  | 'OK'
-  | 'CREATED'
-  | 'ACCEPTED'
-  | 'DELETED'
-  | 'UPDATED';
-
-interface ISuccess<T> {
-  statusCode: SuccessStatusCode;
-  data: T;
-}
-
-export { ISuccess, SuccessStatusCode };
