--- src/@types/statusCodes/index.ts
@@ -1,4 +1,0 @@
-import { IError, ErrorStatusCode } from './IError';
-import { ISuccess, SuccessStatusCode } from './ISuccess';
-
-export { IError, ErrorStatusCode, ISuccess, SuccessStatusCode };
