+++ src/tests/unit/user/create/CreateUserController.spec.ts
@@ -3,7 +3,7 @@ import { User } from '@prisma/client';
 import { expect } from 'chai';
 import Sinon from 'sinon';
 
-import { ISuccess } from '../../../../@types/statusCodes';
+import { ISuccess } from '../../../../@types/interfaces';
 
 import { EncriptService } from '../../../../services/Encript';
 import { UserRepository } from '../../../../modules/users/repository/UsersRepository';
