+++ src/tests/unit/user/create/CreateUserUseCase.spec.ts
@@ -1,7 +1,8 @@
 import { User } from '@prisma/client';
 import { expect } from 'chai';
 import Sinon from 'sinon';
-import { ISuccess } from '../../../../@types/statusCodes';
+
+import { ISuccess } from '../../../../@types/interfaces';
 
 import { UserRepository } from '../../../../modules/users/repository/UsersRepository';
 import { CreateUserUseCase } from '../../../../modules/users/useCases/createUser/CreateUserUseCase';
