+++ src/tests/unit/user/unique/UniqueUserController.spec.ts
@@ -2,7 +2,8 @@ import { NextFunction, request, response } from 'express';
 import { User } from '@prisma/client';
 import { expect } from 'chai';
 import Sinon from 'sinon';
-import { IError, ISuccess } from '../../../../@types/statusCodes';
+
+import { IError, ISuccess } from '../../../../@types/interfaces';
 
 import { UserRepository } from '../../../../modules/users/repository/UsersRepository';
 import { UniqueUserUseCase } from '../../../../modules/users/useCases/uniqueUser/UniqueUserUseCase';
